requirements.txt
asgiref==3.8.1
diff-match-patch==20181111
dj-database-url==2.3.0
Django==5.1.6
django-cors-headers==4.7.0
django-quill-editor==0.1.42
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
fpdf==1.7.2
gunicorn==23.0.0
mysqlclient==2.2.7
ngrok==1.4.0
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pytz==2025.1
quill-delta==1.0.3
sqlparse==0.5.3
telnetlib3==2.0.4
typing_extensions==4.12.2
tzdata==2025.1
